Keep timeline labels fixed while scrolling
